% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfl_cleanForMultiForecast.R
\name{rfl_cleanForMultiForecast}
\alias{rfl_cleanForMultiForecast}
\title{clean data to use rfl_multiForecast()}
\usage{
rfl_cleanForMultiForecast(data, dateCol, activityCol, repNAs = FALSE,
  repSmall = FALSE, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
clean data to use rfl_multiForecast()
}
