% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfl_forecast.R
\name{rfl_forecast}
\alias{rfl_forecast}
\title{Forecast one dataset}
\usage{
rfl_forecast(data, dateCol, activityCol, forecastLength, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Forecast one dataset
}
