% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfl_multiforecast.R
\name{rfl_multiforecast}
\alias{rfl_multiforecast}
\title{Forecast Many datasets}
\usage{
rfl_multiforecast(df_clean, key = "Key", dateCol = "Date",
  activityCol = "Vol", units = "weeks", forecastLength = 52,
  horizon = 10)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Forecast Many datasets
}
